requirements.txt
numpy>=2.2.6
pillow>=11.2.1
sympy>=1.14.0
torch>=2.7.1
torchvision>=0.22.1
triton>=3.3.1
scikit-learn>=1.7.0

